Creating add stock hook
