+++ src/components/forms/AddStocksForm.tsx
@@ -1,13 +1,8 @@
 import { FormEvent, MouseEvent, useEffect, useRef, useState } from 'react';
-import { firebaseClient } from '~/clients/firebase-client/firebase-client';
 import { Stock } from '~/clients/firebase-client/models/Investments';
 import { useAuth } from '~/lib/firebase';
 import DropdownInput from '../shared/DropdownInput';
-import { useDispatch } from 'react-redux';
-import { addStock } from '~/features/investments/investments-slice';
-import getNearestDateRange from '~/helpers/get-nearest-date-range';
-import getBestInterval from '~/helpers/get-best-interval';
-import { addStockData } from '~/features/investments-data/investments-data-slice';
+import useAddStock from '~/hooks/use-add-stock';
 
 export default function AddStocksForm() {
   const [ticker, setTicker] = useState('');
@@ -17,8 +12,7 @@ export default function AddStocksForm() {
   const [activeTab, setActiveTab] = useState<HTMLAnchorElement>();
   const defaultTab = useRef<HTMLAnchorElement>(null);
   const auth = useAuth();
-
-  const dispatch = useDispatch();
+  const addStock = useAddStock();
 
   function handleTabChange(event: MouseEvent<HTMLAnchorElement>) {
     if (event.currentTarget.classList.contains('disabled')) return;
@@ -45,18 +39,8 @@ export default function AddStocksForm() {
         userID: auth.currentUser.uid,
         type: 'stock',
       };
-      firebaseClient().firestore.investments.stocks.add(data);
-      const nearestRange = getNearestDateRange(data.startDate);
-      firebaseClient()
-        .functions.findHistoryStocksBR(
-          [data.ticker],
-          nearestRange,
-          getBestInterval(nearestRange),
-        )
-        .then(res => {
-          dispatch(addStockData(res[0].results[0]));
-          dispatch(addStock(data));
-        });
+
+      addStock(data);
       setTicker('');
       priceInput.current!.value = '';
       amountInput.current!.value = '';
